requirements.txt
opencv-python==4.5.5.62
opencv-contrib-python==4.5.5.62
rainbow-bridge-logger==0.7.0
boto3==1.20.51
fastapi==0.70.1
uvicorn==0.17.4
python-dotenv==0.19.2
mangum==0.12.3
pydantic==1.8.2
requests